# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BookTrainingSessionRequest:
    properties:
      start_time:
        example: "08:00"
        type: string
      trainer_availability_id:
        example: 10
        type: integer
    required:
    - start_time
    - trainer_availability_id
    type: object
  dto.CancelTrainingSessionRequest:
    properties:
      cancellation_reason:
        example: Student has an urgent appointment.
        maxLength: 2000
        minLength: 2
        type: string
    required:
    - cancellation_reason
    type: object
...
  dto.ErrorResponse:
    properties:
      errors:
        type: object
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.RescheduleTrainingSessionRequest:
    properties:
      start_time:
        example: "10:00"
        type: string
      trainer_availability_id:
        example: 11
        type: integer
    required:
    - start_time
    - trainer_availability_id
    type: object
...
  dto.TrainingSessionAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TrainingSessionData'
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.TrainingSessionData:
    properties:
      actual_completed_at:
        type: string
      actual_started_at:
        type: string
      cancellation_reason:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: integer
      created_at:
        type: string
      end_time:
        example: "10:00"
        type: string
      enrollment_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      rescheduled_from_id:
        type: integer
      scheduled_date:
        example: "2026-08-31"
        type: string
      session_number:
        example: 1
        type: integer
      start_time:
        example: "08:00"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.TrainingSessionStatus'
        example: SCHEDULED
      trainer_availability_id:
        example: 10
        type: integer
      trainer_id:
        example: 2
        type: integer
      updated_at:
        type: string
    type: object
  dto.TrainingSessionListAPIResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TrainingSessionData'
        type: array
      message:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.TrainingSessionStatus:
    enum:
    - SCHEDULED
    - IN_PROGRESS
    - COMPLETED
    - CANCELLED
    - RESCHEDULED
    type: string
    x-enum-varnames:
    - SessionScheduled
    - SessionInProgress
    - SessionCompleted
    - SessionCancelled
    - SessionRescheduled
...
info:
  contact: {}
  description: REST API Final Project Bootcamp Intensif Golang
  title: Driving Course Management System API
  version: "1.0"
paths:
...
  /api/v1/admin/training-sessions:
    get:
      description: List all training session history. Requires an ACTIVE ADMIN account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all training session history
      tags:
      - Admin Training Sessions
  /api/v1/admin/training-sessions/{id}:
    get:
      description: Get training session. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get training session
      tags:
      - Admin Training Sessions
  /api/v1/admin/training-sessions/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel a scheduled training session. Requires an ACTIVE ADMIN account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CancelTrainingSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a scheduled training session
      tags:
      - Admin Training Sessions
  /api/v1/admin/training-sessions/{id}/reschedule:
    post:
      consumes:
      - application/json
      description: Reschedule a scheduled training session. Requires an ACTIVE ADMIN
        account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RescheduleTrainingSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reschedule a scheduled training session
      tags:
      - Admin Training Sessions
...
  /api/v1/student/training-sessions:
    get:
      description: List own training session history. Requires an ACTIVE STUDENT account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionListAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List own training session history
      tags:
      - Student Training Sessions
    post:
      consumes:
      - application/json
      description: Book an available two-hour training session. Requires an ACTIVE
        STUDENT account.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BookTrainingSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Book an available two-hour training session
      tags:
      - Student Training Sessions
  /api/v1/student/training-sessions/{id}:
    get:
      description: Get own training session. Requires an ACTIVE STUDENT account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get own training session
      tags:
      - Student Training Sessions
  /api/v1/student/training-sessions/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel own scheduled training session. Requires an ACTIVE STUDENT
        account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CancelTrainingSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel own scheduled training session
      tags:
      - Student Training Sessions
  /api/v1/student/training-sessions/{id}/reschedule:
    post:
      consumes:
      - application/json
      description: Reschedule own scheduled training session. Requires an ACTIVE STUDENT
        account.
      parameters:
      - description: Training session ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RescheduleTrainingSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrainingSessionAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reschedule own scheduled training session
      tags:
      - Student Training Sessions
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter the token with the Bearer prefix, for example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"